pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "peakrdl-viz"
dynamic = ["version"]
requires-python = ">=3.9"
dependencies = [
    "systemrdl-compiler >= 1.29.0, < 2",
    "Jinja2 >= 3.1.2",
]

authors = [
    {name="Ali Mohsen"},
]
description = "Generate visualization code for MakerChip VIZ framework."
readme = "README.md"
license = {text = "LGPLv3"}
keywords = [
    "SystemRDL", "PeakRDL", "CSR", "compiler", "tool", "registers", "generator",
    "MakerChip", "visualization", "VIZ", "Verilog", "SystemVerilog", "TL-Verilog",
    "FPGA", "ASIC",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
]

[project.urls]
Source = "https://github.com/balbal1/PeakRDL_Visualization_Plugin"
Tracker = "https://github.com/balbal1/PeakRDL_Visualization_Plugin/issues"
Changelog = "https://github.com/balbal1/PeakRDL_Visualization_Plugin/releases"

[tool.setuptools.dynamic]
version = {attr = "peakrdl_viz.__about__.__version__"}

[project.entry-points."peakrdl.exporters"]
viz = "peakrdl_viz.__peakrdl__:Exporter"